Refactor + perf utilities:
- Add core/bitops.hpp with select1_index_u64 + random_setbit_index_u64 (BMI2 fast path, portable fallback)
- Refactor PathGraph_64 to use core bit-ops; remove local helpers
- Enable -mbmi/-mbmi2 detection and flags in CMake
- Simplify CliqueGraph to two-color counts-first model; keep CRTP shape and O(1) tf = c1*c2
- Add commented API mirrors in PathGraph_64 for CliqueGraph parity

Perf intent: branch-lean bit selection; BMI2 for throughput, portable fallback otherwise. Complexity: O(1) for selection and clique ops (K=2). No runtime asserts added.

+++ include/core/rng.hpp
@@ -1,4 +1,3 @@
-// rng.hpp
 #pragma once
 #include <array>
 #include <cstdint>
@@ -7,10 +6,20 @@
 
 namespace core {
 
-// SplitMix64 for robust seeding
+/**
+ * @brief SplitMix64 generator for robust seeding.
+ */
 struct SplitMix64 {
     std::uint64_t state;
+    /**
+     * @brief Construct with 64-bit seed.
+     * @param seed Initial state.
+     */
     explicit SplitMix64(std::uint64_t seed) : state(seed) {}
+    /**
+     * @brief Generate next 64-bit value.
+     * @return Next value.
+     */
     inline std::uint64_t next() {
         std::uint64_t z = (state += 0x9E3779B97F4A7C15ULL);
         z = (z ^ (z >> 30)) * 0xBF58476D1CE4E5B9ULL;
@@ -19,27 +28,43 @@ struct SplitMix64 {
     }
 };
 
-// xoshiro256** engine (fast, high-quality for simulations)
+/**
+ * @brief xoshiro256** RNG (fast, high-quality for simulations).
+ */
 struct Xoshiro256ss {
     std::array<std::uint64_t,4> s{0,0,0,0};
-
+    /**
+     * @brief Default constructor (zero state).
+     */
     Xoshiro256ss() = default;
+    /**
+     * @brief Construct and seed from a 64-bit value.
+     * @param seed Seed value.
+     */
     explicit Xoshiro256ss(std::uint64_t seed) { seed_with_u64(seed); }
 
     template <class SeedSeq>
+    /**
+     * @brief Seed the engine from a SeedSeq.
+     * @tparam SeedSeq Seed sequence type with generate(first,last).
+     * @param seq Seed sequence instance.
+     */
     void seed(SeedSeq& seq) {
         std::uint32_t buf[8]{};
         seq.generate(std::begin(buf), std::end(buf));
-        // Pack 8x32 into 4x64
         for (int i=0;i<4;++i) {
             s[i] = (std::uint64_t(buf[2*i]) << 32) | buf[2*i+1];
             if (s[i] == 0) s[i] = 0x9E3779B97F4A7C15ULL ^ (0xBF58476D1CE4E5B9ULL * (i+1));
         }
-        if ((s[0]|s[1]|s[2]|s[3]) == 0) { // all zero is an invalid state
+        if ((s[0]|s[1]|s[2]|s[3]) == 0) {
             seed_with_u64(0xDEADBEEFCAFEBABEULL);
         }
     }
 
+    /**
+     * @brief Seed using SplitMix64 expansion from 64-bit seed.
+     * @param seed Seed value.
+     */
     void seed_with_u64(std::uint64_t seed) {
         SplitMix64 sm(seed);
         for (int i=0;i<4;++i) {
@@ -50,10 +75,20 @@ struct Xoshiro256ss {
         }
     }
 
+    /**
+     * @brief Rotate-left 64-bit value by k bits.
+     * @param x Value to rotate.
+     * @param k Rotation amount.
+     * @return Rotated value.
+     */
     static inline std::uint64_t rotl(const std::uint64_t x, int k) {
         return (x << k) | (x >> (64 - k));
     }
 
+    /**
+     * @brief Generate next 64-bit output.
+     * @return Next 64-bit value.
+     */
     inline std::uint64_t next_u64() {
         const std::uint64_t result = rotl(s[1] * 5ULL, 7) * 9ULL;
         const std::uint64_t t = s[1] << 17;
@@ -69,26 +104,43 @@ struct Xoshiro256ss {
         return result;
     }
 
+    /**
+     * @brief Uniform double in [0,1).
+     * @return Double sample.
+     */
     inline double uniform01() {
-        // 53-bit precision in [0,1)
         return (next_u64() >> 11) * (1.0/9007199254740992.0);
     }
 
+    /**
+     * @brief Uniform integer in [0, bound).
+     * @param bound Upper bound (exclusive).
+     * @return Sampled value without modulo bias.
+     */
     inline std::uint64_t uniform_u64(std::uint64_t bound) {
-        // unbiased modulo reduction
         std::uint64_t x, m = -bound % bound;
         do { x = next_u64(); } while (x < m);
         return x % bound;
     }
 
+    /**
+     * @brief Uniform index in [0, n).
+     * @param n Number of items.
+     * @return Index sample.
+     */
     inline std::size_t uniform_index(std::size_t n) {
-        return static_cast<std::size_t>(uniform_u64(static_cast<std::uint64_t>(n)));
+        return uniform_u64(n);
     }
 };
 
+/**
+ * @brief One-step SplitMix64 hash of a 64-bit value.
+ * @param x Input value.
+ * @return Hashed value.
+ */
 inline std::uint64_t splitmix_hash(std::uint64_t x) {
     SplitMix64 sm(x);
     return sm.next();
 }
 
-} // namespace core
+}
